Options button added
Changed the orientation of the buttons to scale better with the use box containers for organization. Name box container Menu.
Menu: Button addition, option was added in
Option has subdivision of the settings (work in progress). It is divided into Video and Audio.
Video(control node) this has basic implementation to organize what video settings to change with labels for those settings. Settings is a placeholder organizer for how we decide that those video settings should be changed.
Audio(control node): Addition of Labels for settings of Master volume, Music, and soundFX. Slider implementation is basic as there is no audio to text. TODO: added in the section.
All sections have the back button functionality to revert to the main menu.