Test cases for menu
Added test scripts for testing the menu. The addition of a new folder test to store the testscrpits and scene.
Runs test on the function displayHide().
The test then had another function called button_press_test this test all the buttons in the main menu. Does not include the sliders as nothing for sound is set.
Button tested:
_on_options_pressed()
_on_back_options_pressed()
_on_back_video_pressed()
_on_back_audio_pressed()
_on_video_pressed()
_on_audio_pressed()
Test done using asserts