Fixed merge error
Added a bottom GUI for Character, Inventory, Quest, Shop, Settings added the Quest functionality in the game, This is Quest being able to move from In-process to the Completed area. Quest in this menu are scroll-able being that the In-process ones are first and completed shown later. This has a helper window that displays the conditions for completing the quest.
This quest Menu is movable by the player, Note: it does not save any placement in the save file, this could be a feature added. This quest system will remember the quest players have completed since it uses the save file. Files being organized by what system they fall into