Quality of life changes. Quest functional in game
Added many Items with a variety of hats and shoes. Gave the game some responsiveness when window is adjusted but maintain a aspect ratio. Allowed the Inventory to be move by player. Quest in game will now reward the indicated amount to the character be it items, cash, or exp. The exp bar updates and level keep track of if the player reached the threshold to level up this being 100 * level. All this is included in the save file to update everything and boot back everything when game is started again.( possible corruption of old data since many more variables added, resulting in incompatible save file for older version). Added the character to the inventory screen with idle animation.